// frontend/src/shared/api/generated/models/v1WorkDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Jingxuan API
 * OpenAPI spec version: 1.0.0
 */
import type { V1WorkAttachment } from './v1WorkAttachment';
import type { V1WorkMember } from './v1WorkMember';

export interface V1WorkDetail {
  advisor?: string;
  attachments?: V1WorkAttachment[];
  averageScore?: string;
  batchId?: string;
  coverUrl?: string;
  featured?: boolean;
  id?: string;
  likeCount?: number;
  liked?: boolean;
  members?: V1WorkMember[];
  previewUrl?: string;
  publicationStatus?: string;
  rank?: number;
  runDescription?: string;
  status?: string;
  submittedAt?: string;
  submitterId?: string;
  submitterName?: string;
  summary?: string;
  tags?: string[];
  techStack?: string;
  title?: string;
  videoUrl?: string;
  viewCount?: number;
}